Enemy animation and current build
The enemy will now fire at the player when he is facing her and they are
within a range (for testing purposes it is currently a really short
range).  Otherwise he will patrol until he finds the player.
Also added a current build scene for us to keep track of the current
most up to date build.